mv: CgnAnimatedBackground to own file
